animate the reels to move
